# api/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: Varde control plane
  version: "0.2.0"
  description: >
    Public HTTP API for the Varde control plane. The web dashboard uses the
    public endpoints; node agents use the `/v1/agent/*` endpoints (tag `agent`)
    with ed25519 request signing instead of bearer auth.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/nodes/{nodeId}:
    get:
      operationId: getNode
      tags: [nodes]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/NodeId" }]
      responses:
        "200":
          description: Node.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateNode
      tags: [nodes]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/NodeId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                hosting_enabled: { type: boolean }
                anchor: { type: boolean }
                priority: { type: integer, minimum: 0, maximum: 100 }
                max_memory_mb: { type: integer }
                max_cpu_percent: { type: integer }
                max_storage_bytes: { type: integer, format: int64 }
                admin_state:
                  type: string
                  enum: [active, draining, disabled]
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteNode
      tags: [nodes]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/NodeId" }]
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
  /v1/groups/{groupId}/events:
    get:
      operationId: listEvents
      tags: [events]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/GroupId" }
        - name: after
          in: query
          schema: { type: integer, format: int64 }
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
      responses:
        "200":
          description: Events after the given id.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/Event" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /v1/groups/{groupId}/events/stream:
    get:
      operationId: streamEvents
      tags: [events]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/GroupId" }]
      responses:
        "200":
          description: SSE stream; honours Last-Event-ID.
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/groups:
    get:
      operationId: adminListGroups
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        "200":
          description: All groups (operator only).
          content:
            application/json:
              schema:
                type: object
                required: [groups]
                properties:
                  groups:
                    type: array
                    items: { $ref: "#/components/schemas/Group" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/groups/{groupId}/limits:
    put:
      operationId: adminSetGroupLimits
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/GroupId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GroupLimits" }
      responses:
        "200":
          description: Limits set (null = unlimited).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupLimits" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema: { type: string }
...
    NodeId:
      name: nodeId
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: varde_session
    agentSignature:
      type: mutualTLS
      description: >
        Marker scheme — actually ed25519 request signing via headers
        X-Varde-Node, X-Varde-Timestamp, X-Varde-Signature (see
        docs/architecture/control-plane.md §Authentication).
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [stale_epoch, latest_save_unavailable, not_found, forbidden,
                 limit_exceeded, validation, unauthorized, conflict,
                 unschedulable, internal]
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    GroupSettings:
      type: object
      properties:
        snapshot_interval_s: { type: integer, default: 120 }
        default_replication_factor: { type: integer, default: 3 }
        default_min_commit_replicas: { type: integer, default: 2 }
        require_anchor_for_commit: { type: boolean, default: false }
        snapshot_retention: { type: integer, default: 10 }
    Group:
      type: object
      required: [id, name, plan, settings, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        plan: { type: string }
        settings: { $ref: "#/components/schemas/GroupSettings" }
        created_at: { type: integer, format: int64 }
    GroupLimits:
      type: object
      properties:
        max_nodes: { type: integer, nullable: true }
        max_servers: { type: integer, nullable: true }
        max_storage_bytes: { type: integer, format: int64, nullable: true }
        max_relay_bytes_month: { type: integer, format: int64, nullable: true }
...
    Node:
      type: object
      required: [id, group_id, name, public_key, os, arch, hosting_enabled, anchor,
                 admin_state, online, created_at]
      properties:
        id: { type: string }
        group_id: { type: string }
        name: { type: string }
        public_key: { type: string, description: base64 ed25519 }
        os: { type: string }
        arch: { type: string }
        agent_version: { type: string }
        hosting_enabled: { type: boolean }
        anchor: { type: boolean }
        priority: { type: integer }
        max_memory_mb: { type: integer, nullable: true }
        max_cpu_percent: { type: integer, nullable: true }
        max_storage_bytes: { type: integer, format: int64, nullable: true }
        admin_state:
          type: string
          enum: [active, draining, disabled]
        online: { type: boolean, description: "computed from last_seen_at" }
        liveness: { type: string, enum: [online, suspect, offline] }
        last_seen_at: { type: integer, format: int64, nullable: true }
        capabilities: { $ref: "#/components/schemas/Capabilities" }
        created_at: { type: integer, format: int64 }
...
    Event:
      type: object
      required: [id, group_id, type, created_at]
      properties:
        id: { type: integer, format: int64 }
        group_id: { type: string }
        server_id: { type: string, nullable: true }
        node_id: { type: string, nullable: true }
        type: { type: string }
        data: { type: object, additionalProperties: true }
        created_at: { type: integer, format: int64 }

    # ---- agent payloads ----
    Capabilities:
      type: object
      properties:
        os: { type: string }
        arch: { type: string }
        cpu_cores: { type: integer }
        memory_total_mb: { type: integer }
        memory_available_mb: { type: integer }
        disk_free_bytes: { type: integer, format: int64 }
        on_battery: { type: boolean }
        user_active: { type: boolean }
        runtimes:
          type: array
          items: { type: string }
        cached_deployments:
          type: array
          items: { type: string }
        drivers:
          type: array
          items: { type: string }
        uptime_s: { type: integer }
        agent_started_at: { type: integer, format: int64 }
...